Update: directory restructuring
